Update README.md
I added in a sentence to refer to the newest installation instructions to the Wikipedia, which John updated here. (The issue is that the wikipedia is updated, but the README file that people see at the homepage of this github repository was still referring to out of date commands.)

I also added in some back-quote commands since that's the standard way in Markdown language (files that end in ".md") to indicate program names, e.g. `like this`.