Updated analyzer script so now it correctly does log reg.
In fact, I think our paper has something weird. How are we getting -25 for log likelihood? With binary classfication, if we say probability of 1/2 in many locations, that would imply we're getting log(1/2) or -0.69, which makes sense. But our value is much larger. The only alternative I can think of why is if we randomliy initialized and got maybe log(0.000000001) somewhere which skewed the average, but even so we'd need something REALLY small ...